requirements.txt
aniso8601==9.0.1
attrs==23.1.0
beautifulsoup4==4.9.3
blinker==1.6.3
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
cssselect==1.2.0
Faker==19.3.1
Flask==2.1.0
flask-restplus==0.13.0
flask-restx==0.5.1
gevent==23.9.0.post1
greenlet==3.0.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
lxml==4.9.3
MarkupSafe==2.1.3
pyquery==2.0.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
requests==2.30.0
rpds-py==0.10.6
six==1.16.0
soupsieve==2.5
urllib3==2.0.6
Werkzeug==2.1.2
zope.event==5.0
zope.interface==6.1
